Create the Estate List and Estate Detail fragments with fake data, and manage the layout for both tablet and mobile.

+++ app/src/main/java/com/openclassrooms/realestatemanager/data/models/Photo.kt
@@ -1,3 +1,3 @@
 package com.openclassrooms.realestatemanager.data.models
 
-data class Photo(val id: Int, val photoPath: String, val description: String)
+data class Photo(val id: Long, val photoPath: String, val description: String)
